Improve usage of types
